+++ renderer/style/message.module.css
@@ -1,0 +1,11 @@
+.messageCover {
+    height: 80px;
+    border-radius: 20px;
+    margin: 5px 10px;
+    transition: 0.2s;
+    cursor: pointer;
+}
+
+.messageCover:hover {
+    background-color: #a5a8ac;
+}
