+++ renderer/tsconfig.json
@@ -4,7 +4,7 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx"
-  ],
+, "pages/chatList.jsx"  ],
   "exclude": [
     "node_modules"
   ],
